% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senso.R
\docType{package}
\name{senso}
\alias{senso-package}
\alias{senso}
\title{An Interface to the Philippine Census of Population and Housing Data}
\description{
The Philippine Statistics Authority (PSA) makes census of population and
housing data publicly available through various data publications and
releases over time via their website. This package provides utilities for
accessing and processing the census of population and housing data released
by the PSA.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
